Share event configuration between create and edit
