+++ app-modules/meeting-center/src/Filament/Resources/EventResource/Pages/CreateEvent.php
@@ -36,43 +36,19 @@
 
 namespace AdvisingApp\MeetingCenter\Filament\Resources\EventResource\Pages;
 
-use App\Models\User;
 use Filament\Forms\Form;
-use Filament\Forms\Components\Textarea;
-use Filament\Forms\Components\TextInput;
 use Filament\Resources\Pages\CreateRecord;
-use Filament\Forms\Components\DateTimePicker;
 use AdvisingApp\MeetingCenter\Filament\Resources\EventResource;
+use AdvisingApp\MeetingCenter\Filament\Resources\EventResource\Pages\Concerns\HasSharedEventFormConfiguration;
 
 class CreateEvent extends CreateRecord
 {
+    use HasSharedEventFormConfiguration;
+
     protected static string $resource = EventResource::class;
 
     public function form(Form $form): Form
     {
-        /** @var User $user */
-        $user = auth()->user();
-
-        return $form->schema([
-            TextInput::make('title')
-                ->string()
-                ->required(),
-            Textarea::make('description')
-                ->string()
-                ->nullable(),
-            TextInput::make('location')
-                ->string()
-                ->nullable(),
-            TextInput::make('capacity')
-                ->integer()
-                ->minValue(1)
-                ->nullable(),
-            DateTimePicker::make('starts_at')
-                ->timezone($user->timezone)
-                ->required(),
-            DateTimePicker::make('ends_at')
-                ->timezone($user->timezone)
-                ->required(),
-        ]);
+        return $form->schema($this->fields());
     }
 }
